+++ classes/Builders.py
@@ -12,29 +12,6 @@ class Builder:
     def __init__(self):
         self.db: Database = Database()
 
-    def build_ratings(self):
-        ratings_data = []
-
-        with open("data/ratings.txt", encoding="utf8") as f:
-            next(f) # skip header line
-
-            for line in f:
-                rating, question_id, title, _, title_slug, _, _ = line.split("\t")
-
-                ratings_structure = {
-                    "question_id": int(question_id),
-                    "title": title,
-                    "title_slug": title_slug,
-                    "rating": float(rating)
-                }
-
-                ratings_data.append(ratings_structure)
-
-        self.db.delete_ratings_db()
-        self.db.insert_ratings_db(ratings_data)
-
-        print(f"Successfully pushed questions ratings from ratings.txt to database")
-
     def build_question_rating_data(self):
         # now that we have question bank, we can simplify all of the questions
         # to build association between rating, question, and tags. since we are
@@ -51,7 +28,7 @@ class Builder:
             },
             {
                 '$addFields': {
-                    'question_id': {'$toInt': '$questionFrontendId'}
+                    'question_id': '$questionFrontendId'
                 }
             },
             {
@@ -60,17 +37,6 @@ class Builder:
                     'localField': 'question_id',
                     'foreignField': 'question_id',
                     'as': 'joined_docs'
-                }
-            },
-            {
-                '$addFields': {
-                    'joined_docs': {
-                        '$cond': {
-                            'if': {'$eq': ['$joined_docs', []]},  # Check if joined_docs is empty
-                            'then': [{'rating': '$predicted_rating'}],  # If empty, set rating to predicted_rating from model
-                            'else': '$joined_docs'  # Otherwise, keep it as is
-                        }
-                    }
                 }
             },
             {
@@ -106,7 +72,8 @@ class Builder:
                     "title": "$questionTitle",
                     "title_slug": "$titleSlug",
                     "link": "$link",
-                    "rating": "$rating",
+                    "predicted_rating": "$predicted_rating",
+                    "zerotrac_rating": "$zerotrac_rating",
                     "difficulty": "$difficulty",
                     "total_acs": "$total_acs",
                     "total_submitted": "$total_submitted",
@@ -131,11 +98,6 @@ class Builder:
 
         user_data_structure = user.output_user_data_structure()
 
-        self.db.user_data_collection.insert_one(user_data_structure)
+        self.db.insert_user(user_data_structure)
 
         print(f"Successfully built user database for user {user_discord_username}")
-
-    def check_user_exist(self, user_id: int):
-        result = self.db.user_data_collection.find_one({"discord_id": user_id})
-
-        return result != None