+++ classes/Parsers.py
@@ -9,29 +9,9 @@ class Parser:
     def __init__(self):
         self.db: Database = Database()
 
-    def parse_singular_question(self, title_slug):
-        # util function for just calling the api once on a single question, for maybe
-        # when the api throws an error and gives bad response
-        
-        with open("data/user.txt", "r+") as f:
-            user_data = json.load(f)
-
-        r = api_get_question_info(title_slug)
-
-        question = r.json()
-        question["questionId"] = int(question["questionId"])
-        question["questionId"] = int(question["questionFrontendId"])
-        # TODO: fix these
-        # question["total_acs"] = total_acs
-        # question["total_submitted"] = total_submitted
-
-        self.db.insert_question(question)
-
     def parse_question_bank(self):
         with open("data/user.txt", "r+") as f:
             user_data = json.load(f)
-
-        self.db.delete_questions_db()
 
         all_questions = user_data["stat_status_pairs"]
 
@@ -41,10 +21,12 @@ class Parser:
             total_acs = question["stat"]["total_acs"]
             total_submitted = question["stat"]["total_submitted"]
 
-            question = self.db.find_question(question_id)
+            question = self.db.find_question_by_question_id(question_id)
 
             if question:
                 print(f"question id {question_id} parsed already, skipping")
+
+                time.sleep(0.1 * 1) # so that we don't hit cluster with too many req/s
                 continue
 
             r = api_get_question_info(title_slug)
@@ -61,10 +43,41 @@ class Parser:
 
             question = r.json()
             question["questionId"] = int(question["questionId"])
-            question["questionId"] = int(question["questionFrontendId"])
+            question["questionFrontendId"] = int(question["questionFrontendId"])
             question["total_acs"] = total_acs
             question["total_submitted"] = total_submitted
 
             self.db.insert_question(question)
 
             time.sleep(10 * 1) # 10 seconds between each call
+
+    def build_ratings(self):
+        # this method will take some time to run, as we are predicting using
+        # the model on many questions
+
+        self.db.delete_ratings_db()
+
+        # preload ratings into a dictionary first
+        ratings = dict()
+
+        with open("data/ratings.txt", encoding="utf8") as f:
+            next(f) # skip header line
+
+            for line in f:
+                rating, question_id, _, _, _, _, _ = line.split("\t")
+
+                ratings[int(question_id)] = float(rating)
+
+        questions = self.db.return_all_raw_questions()
+
+        ratings_data = [{
+            "question_id": question["questionFrontendId"],
+            "title": question["questionTitle"],
+            "title_slug": question["titleSlug"],
+            "predicted_rating": predict_question_rating(question),
+            "zerotrac_rating": ratings.get(question["questionFrontendId"], 0)
+        } for question in questions]
+
+        self.db.insert_ratings_db(ratings_data)
+
+        print(f"Successfully pushed questions ratings to database")