+++ main.py
@@ -9,10 +9,10 @@ def main():
     # builder = Builder()
     # builder.build_question_rating_data()
 
-    parser = Parser()
-    parser.parse_question_bank()
+    # parser = Parser()
+    # parser.build_ratings()
 
-    # setup_bot()
+    setup_bot()
 
 def setup_bot():
     intents = discord.Intents.default()